Move benchmarks under benches/ so tests compile on stable
Closes #36.
