+++ src/bls12_381/fq.rs
@@ -1138,107 +1138,6 @@ fn test_fq_repr_add_nocarry() {
     assert!(!x.add_nocarry(&FqRepr::from(1)));
 }
 
-#[bench]
-fn bench_fq_repr_add_nocarry(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<(FqRepr, FqRepr)> = (0..SAMPLES).map(|_| {
-        let mut tmp1 = FqRepr::rand(&mut rng);
-        let mut tmp2 = FqRepr::rand(&mut rng);
-        // Shave a few bits off to avoid overflow.
-        for _ in 0..3 {
-            tmp1.div2();
-            tmp2.div2();
-        }
-        (tmp1, tmp2)
-    }).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count].0;
-        tmp.add_nocarry(&v[count].1);
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
-#[bench]
-fn bench_fq_repr_sub_noborrow(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<(FqRepr, FqRepr)> = (0..SAMPLES).map(|_| {
-        let tmp1 = FqRepr::rand(&mut rng);
-        let mut tmp2 = tmp1;
-        // Ensure tmp2 is smaller than tmp1.
-        for _ in 0..10 {
-            tmp2.div2();
-        }
-        (tmp1, tmp2)
-    }).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count].0;
-        tmp.sub_noborrow(&v[count].1);
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
-#[bench]
-fn bench_fq_repr_num_bits(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<FqRepr> = (0..SAMPLES).map(|_| FqRepr::rand(&mut rng)).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let tmp = v[count].num_bits();
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
-#[bench]
-fn bench_fq_repr_mul2(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<FqRepr> = (0..SAMPLES).map(|_| FqRepr::rand(&mut rng)).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count];
-        tmp.mul2();
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
-#[bench]
-fn bench_fq_repr_div2(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<FqRepr> = (0..SAMPLES).map(|_| FqRepr::rand(&mut rng)).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count];
-        tmp.div2();
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
 #[test]
 fn test_fq_is_valid() {
     let mut a = Fq(MODULUS);
@@ -1350,40 +1249,6 @@ fn test_fq_sub_assign() {
     }
 }
 
-#[bench]
-fn bench_fq_add_assign(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<(Fq, Fq)> = (0..SAMPLES).map(|_| (Fq::rand(&mut rng), Fq::rand(&mut rng))).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count].0;
-        tmp.add_assign(&v[count].1);
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
-#[bench]
-fn bench_fq_sub_assign(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<(Fq, Fq)> = (0..SAMPLES).map(|_| (Fq::rand(&mut rng), Fq::rand(&mut rng))).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count].0;
-        tmp.sub_assign(&v[count].1);
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
 #[test]
 fn test_fq_mul_assign() {
     let mut tmp = Fq(FqRepr([0xcc6200000020aa8a, 0x422800801dd8001a, 0x7f4f5e619041c62c, 0x8a55171ac70ed2ba, 0x3f69cc3a3d07d58b, 0xb972455fd09b8ef]));
@@ -1433,23 +1298,6 @@ fn test_fq_mul_assign() {
     }
 }
 
-#[bench]
-fn bench_fq_mul_assign(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<(Fq, Fq)> = (0..SAMPLES).map(|_| (Fq::rand(&mut rng), Fq::rand(&mut rng))).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count].0;
-        tmp.mul_assign(&v[count].1);
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
 #[test]
 fn test_fq_squaring() {
     let mut a = Fq(FqRepr([0xffffffffffffffff, 0xffffffffffffffff, 0xffffffffffffffff, 0xffffffffffffffff, 0xffffffffffffffff, 0x19ffffffffffffff]));
@@ -1473,23 +1321,6 @@ fn test_fq_squaring() {
     }
 }
 
-#[bench]
-fn bench_fq_square(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<Fq> = (0..SAMPLES).map(|_| Fq::rand(&mut rng)).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count];
-        tmp.square();
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
 #[test]
 fn test_fq_inverse() {
     assert!(Fq::zero().inverse().is_none());
@@ -1505,21 +1336,6 @@ fn test_fq_inverse() {
         a.mul_assign(&ainv);
         assert_eq!(a, one);
     }
-}
-
-#[bench]
-fn bench_fq_inverse(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<Fq> = (0..SAMPLES).map(|_| Fq::rand(&mut rng)).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        count = (count + 1) % SAMPLES;
-        v[count].inverse()
-    });
 }
 
 #[test]
@@ -1557,24 +1373,6 @@ fn test_fq_negate() {
         assert!(a.is_zero());
     }
 }
-
-#[bench]
-fn bench_fq_negate(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<Fq> = (0..SAMPLES).map(|_| Fq::rand(&mut rng)).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        let mut tmp = v[count];
-        tmp.negate();
-        count = (count + 1) % SAMPLES;
-        tmp
-    });
-}
-
 
 #[test]
 fn test_fq_pow() {
@@ -1631,25 +1429,6 @@ fn test_fq_sqrt() {
     }
 }
 
-#[bench]
-fn bench_fq_sqrt(b: &mut ::test::Bencher) {
-	const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<Fq> = (0..SAMPLES).map(|_| {
-        let mut tmp = Fq::rand(&mut rng);
-        tmp.square();
-        tmp
-    }).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        count = (count + 1) % SAMPLES;
-        v[count].sqrt()
-    });
-}
-
 #[test]
 fn test_fq_from_into_repr() {
     // q + 1 should not be in the field
@@ -1682,40 +1461,6 @@ fn test_fq_from_into_repr() {
 
         assert_eq!(a, a_again);
     }
-}
-
-#[bench]
-fn bench_fq_into_repr(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<Fq> = (0..SAMPLES).map(|_| {
-        Fq::rand(&mut rng)
-    }).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        count = (count + 1) % SAMPLES;
-        v[count].into_repr()
-    });
-}
-
-#[bench]
-fn bench_fq_from_repr(b: &mut ::test::Bencher) {
-    const SAMPLES: usize = 1000;
-
-    let mut rng = XorShiftRng::from_seed([0x5dbe6259, 0x8d313d76, 0x3237db17, 0xe5bc0654]);
-
-    let v: Vec<FqRepr> = (0..SAMPLES).map(|_| {
-        Fq::rand(&mut rng).into_repr()
-    }).collect();
-
-    let mut count = 0;
-    b.iter(|| {
-        count = (count + 1) % SAMPLES;
-        Fq::from_repr(v[count])
-    });
 }
 
 #[test]
